# Reuse hash in CombineTraceProtos to reduce memory allocations