@joe-elliott 2020-12-02
Nice improvement.  Probably should do something similar in `getOrCreateTrace()` in instance.go.  This line is hit 1000s of times per second:

https://github.com/grafana/tempo/blob/c0cb2fabb58edd55184ee190c9f4cf19d42e7f18/modules/ingester/instance.go#L279